Reorganized Interface tests
