+++ tests/core/data/testirisinterface.py
@@ -1,21 +1,98 @@
+from nose.plugins.attrib import attr
+from unittest import SkipTest
+
 import numpy as np
-import unittest
 
 try:
     from iris.tests.stock import lat_lon_cube
 except ImportError:
-    raise unittest.SkipTest("Could not import iris, skipping iris interface "
-                            "tests.")
+    raise SkipTest("Could not import iris, skipping IrisInterface tests.")
 
 from holoviews.core.data import Dataset
 from holoviews.core.data.iris import coord_to_dimension
-from holoviews.element.comparison import ComparisonTestCase
+from holoviews.element import Image
 
-class TestCube(ComparisonTestCase):
+from .testimageinterface import Image_ImageInterfaceTests
+from .testgridinterface import GridInterfaceTests
+
+
+@attr(optional=1)
+class IrisInterfaceTests(GridInterfaceTests):
+    """
+    Tests for Iris interface
+    """
+
+    datatype = 'cube'
 
     def setUp(self):
+        import iris
+        self.restore_datatype = Dataset.datatype
+        Dataset.datatype = ['cube']
+        self.eltype = Dataset
+        self.data_instance_type = iris.cube.Cube
+        self.init_column_data()
+        self.init_grid_data()
         self.cube = lat_lon_cube()
         self.epsilon = 0.01
+
+    def test_dataset_array_init_hm(self):
+        "Tests support for arrays (homogeneous)"
+        raise SkipTest("Not supported")
+
+    # Disabled tests for NotImplemented methods
+    def test_dataset_add_dimensions_values_hm(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_add_dimensions_values_hm_alias(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_sort_hm(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_sort_reverse_hm(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_sort_vdim_hm(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_sort_reverse_vdim_hm(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_sort_vdim_hm_alias(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_1D_reduce_hm(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_1D_reduce_hm_alias(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_2D_reduce_hm(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_2D_reduce_hm_alias(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_2D_aggregate_partial_hm(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_2D_aggregate_partial_hm_alias(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_sample_hm(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_sample_hm_alias(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_groupby_drop_dims_with_vdim(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_groupby_drop_dims_dynamic_with_vdim(self):
+        raise SkipTest("Not supported")
+
+    def test_dataset_ndloc_slice_two_vdims(self):
+        raise SkipTest("Not supported")
 
     def test_dim_to_coord(self):
         dim = coord_to_dimension(self.cube.coords()[0])
@@ -103,3 +180,72 @@ class TestCube(ComparisonTestCase):
     def test_getitem_scalar(self):
         cube = Dataset(self.cube)
         self.assertEqual(cube[0, 0], 5)
+
+
+@attr(optional=1)
+class Image_IrisInterfaceTests(Image_ImageInterfaceTests):
+
+    datatype = 'cube'
+
+    def init_data(self):
+        xs = np.linspace(-9, 9, 10)
+        ys = np.linspace(0.5, 9.5, 10)
+        self.xs = xs
+        self.ys = ys
+        self.array = np.arange(10) * np.arange(10)[:, np.newaxis]
+        self.image = Image((xs, ys, self.array))
+        self.image_inv = Image((xs[::-1], ys[::-1], self.array[::-1, ::-1]))
+
+    def test_init_data_datetime_xaxis(self):
+        raise SkipTest("Not supported")
+
+    def test_init_data_datetime_yaxis(self):
+        raise SkipTest("Not supported")
+
+    def test_init_bounds_datetime_xaxis(self):
+        raise SkipTest("Not supported")
+
+    def test_init_bounds_datetime_yaxis(self):
+        raise SkipTest("Not supported")
+
+    def test_init_densities_datetime_xaxis(self):
+        raise SkipTest("Not supported")
+
+    def test_init_densities_datetime_yaxis(self):
+        raise SkipTest("Not supported")
+
+    def test_range_datetime_xdim(self):
+        raise SkipTest("Not supported")
+
+    def test_range_datetime_ydim(self):
+        raise SkipTest("Not supported")
+
+    def test_dimension_values_datetime_xcoords(self):
+        raise SkipTest("Not supported")
+
+    def test_dimension_values_datetime_ycoords(self):
+        raise SkipTest("Not supported")
+
+    def test_slice_datetime_xaxis(self):
+        raise SkipTest("Not supported")
+
+    def test_slice_datetime_yaxis(self):
+        raise SkipTest("Not supported")
+
+    def test_reduce_to_scalar(self):
+        raise SkipTest("Not supported")
+
+    def test_reduce_x_dimension(self):
+        raise SkipTest("Not supported")
+
+    def test_reduce_y_dimension(self):
+        raise SkipTest("Not supported")
+
+    def test_aggregate_with_spreadfn(self):
+        raise SkipTest("Not supported")
+
+    def test_sample_datetime_xaxis(self):
+        raise SkipTest("Not supported")
+
+    def test_sample_datetime_yaxis(self):
+        raise SkipTest("Not supported")
